chore(clippy): make workspace warning-free
- Fix remaining clippy lints across workspace (tests/benches included) so `cargo clippy --workspace --all-targets --all-features -- -D warnings` passes.
- Update emulator network integration tests to the current nt-packetlib/aero_net_stack packet builder APIs.
- Replace manual modulo/alignment checks with `.is_multiple_of()`, remove identity/erasing ops, and prefer struct update literals over default+reassign patterns where applicable.

+++ tests/d3d9_blend_depth_stencil.rs
@@ -251,9 +251,9 @@ fn assert_rgba_near(actual: [u8; 4], expected: [u8; 4], tolerance: u8) {
     for i in 0..4 {
         let a = actual[i] as i16;
         let e = expected[i] as i16;
-        let diff = (a - e).unsigned_abs() as u8;
+        let diff = (a - e).unsigned_abs();
         assert!(
-            diff <= tolerance,
+            diff <= u16::from(tolerance),
             "channel {}: actual={:?} expected={:?} tolerance={}",
             i,
             actual,
